+++ docs/release-notes.md
@@ -2,6 +2,7 @@
 - Shorten long anime titles in progress bars
 - Hover on folder button shows folder location
 - senpcli: Always show units/time instead of time/units in progress bars
+- senpcli: Add anime tracking functionality
 
 # Bug fixes
 - Fix some gogo downloads randomly failing, if you still experience it be sure to report it